* Barco Asym support (#1490)
* Functional ECC Point Mul for P256
* Move the TLS header scratch in Post-Encrypt PHV to above DMA commands

+++ nic/hal/pd/capri/capri_barco_res.hpp
@@ -13,7 +13,7 @@ typedef struct capri_barco_resources_s {
     uint16_t                    obj_size;
     uint16_t                    obj_alignment;
     uint64_t                    hbm_region;
-    uint16_t                    hbm_region_size;
+    uint32_t                    hbm_region_size;
     indexer                     *idxer;
 } capri_barco_resources_t;
